Refactor cloud images build code
Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1659
Tested-by: gerrit-photon <photon-checkins@vmware.com>

Do not delete existing ova artifacts and additional fixes

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1663
Tested-by: suezzelur <anishs@vmware.com>

Rename vmx template files and change default tar format

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1666
Tested-by: suezzelur <anishs@vmware.com>

Change tar format for ami,gce

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1671
Tested-by: suezzelur <anishs@vmware.com>

Remove unneeded files, reduce the size of ami raw image

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1673
Tested-by: suezzelur <anishs@vmware.com>

Change vmdk size of ovas

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1683
Tested-by: suezzelur <anishs@vmware.com>

Change grub for ami,gce to accept photon.cfg values

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1691
Tested-by: suezzelur <anishs@vmware.com>

Edit fstab to include UUID value of disk for ami, gce

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1692
Tested-by: suezzelur <anishs@vmware.com>

Change kernel parameters for ami image

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1698
Tested-by: suezzelur <anishs@vmware.com>

cloud_image - PARTUUID in lowered case

systemd-remount-fs.service require PARTUUIS in lowered case.

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/1693
Tested-by: gerrit-photon <photon-checkins@vmware.com>

Remove mk-setup-grub for cloud images

+++ support/cloud-image-builder/ami/ami-patch.sh
@@ -8,8 +8,11 @@ cd /
 echo "127.0.0.1 localhost" >> /etc/hosts
 
 # Update /etc/resolv.conf
-rm /etc/resolv.conf
-echo "nameserver 172.31.0.2" >> /etc/resolv.conf
+if [ -f /etc/resolv.conf ]
+	then
+		rm /etc/resolv.conf
+fi
+echo "nameserver 169.254.169.253" >> /etc/resolv.conf
 echo "search ec2.internal" >> /etc/resolv.conf
 
 
@@ -54,11 +57,14 @@ echo "Ciphers aes128-ctr,aes192-ctr,aes256-ctr,arcfour256,arcfour128,aes128-cbc,
 echo "Tunnel no" >> /etc/ssh/ssh_config
 echo "ServerAliveInterval 420" >> /etc/ssh/ssh_config
 
-sed -i '/.*linux.*vmlinuz/ s/$/ console=ttyS0/' /boot/grub/grub.cfg
+sed -i 's/net.ifnames=0//' /boot/grub/grub.cfg
+sed -i 's/$photon_cmdline/init=\/lib\/systemd\/systemd loglevel=3 ro console=ttyS0 earlyprintk=ttyS0/' /boot/grub/grub.cfg
 
 # Disable loading/unloading of modules
 echo 1 > /proc/sys/kernel/modules_disabled
 
 # Remove kernel symbols
-rm /boot/system.map*
-
+if [ -f /boot/system.map* ]
+	then
+		rm /boot/system.map*
+fi
